Updated childContracts to behave like the base parent data service with a slight difference. For the singleton, any resource view below it will be immediately transformed to a singleton, since it's a view of a singleton. We can just construct the endpoint as this.endpoint + contract.endpoint, since the parent singleton should already be scoped against the correct entity. (Example - /serviceEvent/9008/performedBy/child)
For parent resource views, this means that for a child that is defined as a view of the parent view, the consumer will be able to hit /serviceEvent/performedBy/child as a simple resource or /serviceEvent/9008/performedBy/child as a singleton.

The only minor issue here is that mock data will be handled incorrectly if a resource view is defined as a direct child of a singleton (not a view). Since the parent only gets an id upon transformation as a view, a parent that is defined as a singleton will implicitly expect all children to be singletons or collections.