Reimplement tests for synchronizedRequests.service.tests
We couldn't get this to work before because the assertions would get hit before the promise was resolved. We can get around this by doing the asserts in a .then() after the flush call. Also use a new async helper similar to the one for jasmine from angular 2 to run the test in a zone and ensure that the test doesn't exit until after all promises have resolved.