Throw an error if requests are pending
We want to have a way for fakeAsync to fail the test if there are still pending requests. We'll do this by having mockPromise plug in and queue a request with pending set to true everytime a promise is constructed. When flushed, pending is cleared. After the test code finishes, we'll check the queued requests. If any are still pending, we'll throw an exception.