Create HttpDigestService
HTTP requests were using the ng2 http module, which returns an observable and was being converted to a promise.  Because of this, the digest cycle was not firing and elements in the DOM were not updating.

Created an ng2 HTTP interceptor service that will explicitly run a digest cycle after HTTP requests succeed or fail.  Because scope does not exist in ng2, the interceptor was designed to accept scope.applyAsync as a parameter and call it.