Refactor HttpDigestService into DigestService
The logic that obtains a reference to the scope and calls `$applyAsync` has been moved out of Application.Core and into this service.

The functionality of this service is not tied to HTTP in any way and it is reusable, so it was renamed from `HttpDigestService` to just `DigestService`.

The service takes no dependencies and exposes just one function that takes no parameters and returns nothing.  Unsure how to test it or whether it needs to be tested.

Note: This will die when ng1 is dead