isNullOrEmpty should return false for [0]
We're having an issue where isNullOrEmpty is returning true for arrays that have items but the items are falsy. We'll fix this by using a length check instead of _.some. I'll have to look out for that in the future. It's probably best to only use some with a predicate in most cases, because of this edge case.