Make isEmpty impure
The concept of a 'pure' pipe is one in which the output only changes if the input changes. Apparently, this is affected by the mutable/immutable question. A pure pipe will only update if the reference of its dependency changes, and thus is broken by mutability. By setting the pipe to impure, it picks up changes properly with mutable data structures.