Remove opaque tokens
As a team we decided we like the simplicity of specifying a class without having to use the token. As much as possible, I'm moving toward using classes and abstract classes where we would've used a token before. Also changed a few lets to consts and renamed module files to index.ts