Added a helper for mocking data contracts
Instead of taking the approach of allowing application code to create and use mock contracts, we'll simply provide a helper function that sets the mock data on the contract and specifies that useMock should be true. This still allows the application to pull the mock data out of the contract definition and into the prototype application code, while also still integrating with the logging capabilities built into data contracts. The only drawback to this approach is that a data contract, once set to use mocks, will use mocks everywhere it is used, unless it is first cloned.