Allow for scheduling rxjs tasks with fakeAsync
After some investigation, I've discovered that the reason scheduling rxjs tests is currently not possible is that the scheduler always depends on the current browser time. If we instead tie this to the elapsed time of the fake async zone based on 'tick', we'll gain the ability to also schedule rxjs tasks in tests. Requires wrapping angular 'tick' so we can track the elapsed time.