Need systemjs as an explicit dependency
We'll just add it as a dev dependency since we don't actually care that our consumers have it when they run the app. (They just need to use our javascript files)

This is the version referenced by app core. In components, we use the '^' so I'm going to be equally flexible here.