Make failing unit tests use at least comparison instead of equals
Because moment was still ticking on when this test was running, a direct equality comparison would sometimes fail. Moment doesn't seem to have any methods to freeze the date time either.
This seems to be the best way to write the test. Since we're adding the offset after "nowBeforeOffset" is initialized, "nowAfterOffset" should always be at least the offset. If it's any less, then we know something went very wrong.
Also, change the offset to be an unholy amount (about 5 days now). That way, we don't mistake the passage of time for a scenario that would let the test pass.