Install @types/jquery@2.0.32
Angular screws us over by installing the latest version of @types/jquery. It took me a bit, but it looks like it's because they set a dependency on version `*` of jquery. It works in components because we have an explicit dependency set. Setting this explicitly here seems to have fixed it.